feat: 新增ESP32-C3支持和完整文档 - 新增ESP32-C3开发板支持(合宙CORE-ESP32-C3) - 实现直连4.2寸电子墨水屏，无需驱动板 - 新增开机自动屏幕测试功能 - 支持板载BOOT按钮(GPIO9)和LED(GPIO12) - 优化接线方案，使用连续引脚17-24 - 新增构建脚本和完整文档 - 修复ESP32-C3兼容性问题 - 感谢原作者JADE-Jerry的开源贡献

+++ src/API.hpp
@@ -106,7 +106,7 @@ private:
                     if (isGzip) {
                         // gzip解压缩
                         uint8_t* outBuf = NULL;
-                        size_t outLen = 0;
+                        uint32_t outLen = 0;  // 修改为uint32_t类型
                         ArduinoUZlib::decompress((uint8_t*)s.c_str(), (uint32_t)s.length(), outBuf, outLen);
                         error = deserializeJson(doc, (char*)outBuf, outLen);
                     } else {
